+++ src/main/java/com/project/bookstudy/post/service/PostService.java
@@ -7,10 +7,10 @@ import com.project.bookstudy.common.exception.FileException;
 import com.project.bookstudy.member.domain.Member;
 import com.project.bookstudy.member.repository.MemberRepository;
 import com.project.bookstudy.post.domain.Post;
-import com.project.bookstudy.post.dto.CreatePostRequest;
-import com.project.bookstudy.post.dto.CreatePostResponse;
+import com.project.bookstudy.post.dto.request.CreatePostRequest;
+import com.project.bookstudy.post.dto.response.CreatePostResponse;
 import com.project.bookstudy.post.dto.PostDto;
-import com.project.bookstudy.post.dto.PostSearchCond;
+import com.project.bookstudy.post.dto.request.PostSearchCond;
 import com.project.bookstudy.post.file.domain.File;
 import com.project.bookstudy.post.file.repository.FileRepository;
 import com.project.bookstudy.post.file.service.S3Deleter;
